+++ sdk/application/Cart/CartApplicationValidator.d.ts
@@ -76,7 +76,6 @@ export = CartApplicationValidator;
  * @property {string} [id]
  * @property {boolean} [i]
  * @property {boolean} [b]
- * @property {boolean} [c]
  * @property {number} [assignCardId]
  * @property {string} [areaCode]
  * @property {boolean} [buyNow]
@@ -346,7 +345,6 @@ type GetCartParam = {
     id?: string;
     i?: boolean;
     b?: boolean;
-    c?: boolean;
     assignCardId?: number;
     areaCode?: string;
     buyNow?: boolean;