+++ sdk/application/Cart/CartApplicationValidator.js
@@ -89,7 +89,6 @@ const CartApplicationModel = require("./CartApplicationModel");
  * @property {string} [id]
  * @property {boolean} [i]
  * @property {boolean} [b]
- * @property {boolean} [c]
  * @property {number} [assignCardId]
  * @property {string} [areaCode]
  * @property {boolean} [buyNow]
@@ -336,7 +335,6 @@ class CartApplicationValidator {
       id: Joi.string().allow(""),
       i: Joi.boolean(),
       b: Joi.boolean(),
-      c: Joi.boolean(),
       assignCardId: Joi.number(),
       areaCode: Joi.string().allow(""),
       buyNow: Joi.boolean(),