+++ sdk/application/PosCart/PosCartApplicationClient.js
@@ -832,13 +832,13 @@ class PosCart {
    * @description: Use this API to get details of all the items added to a cart. - Check out [method documentation](https://partners.fynd.com/help/docs/sdk/application/poscart/getCart/).
    */
   async getCart(
-    { id, i, b, c, assignCardId, areaCode, buyNow, requestHeaders } = {
+    { id, i, b, assignCardId, areaCode, buyNow, requestHeaders } = {
       requestHeaders: {},
     },
     { responseHeaders } = { responseHeaders: false }
   ) {
     const { error } = PosCartApplicationValidator.getCart().validate(
-      { id, i, b, c, assignCardId, areaCode, buyNow },
+      { id, i, b, assignCardId, areaCode, buyNow },
       { abortEarly: false, allowUnknown: true }
     );
     if (error) {
@@ -847,7 +847,7 @@ class PosCart {
 
     // Showing warrnings if extra unknown parameters are found
     const { error: warrning } = PosCartApplicationValidator.getCart().validate(
-      { id, i, b, c, assignCardId, areaCode, buyNow },
+      { id, i, b, assignCardId, areaCode, buyNow },
       { abortEarly: false, allowUnknown: false }
     );
     if (warrning) {
@@ -861,7 +861,6 @@ class PosCart {
     query_params["id"] = id;
     query_params["i"] = i;
     query_params["b"] = b;
-    query_params["c"] = c;
     query_params["assign_card_id"] = assignCardId;
     query_params["area_code"] = areaCode;
     query_params["buy_now"] = buyNow;