+++ sdk/application/PosCart/PosCartApplicationValidator.d.ts
@@ -72,7 +72,6 @@ export = PosCartApplicationValidator;
  * @property {string} [id]
  * @property {boolean} [i]
  * @property {boolean} [b]
- * @property {boolean} [c]
  * @property {number} [assignCardId]
  * @property {string} [areaCode]
  * @property {boolean} [buyNow]
@@ -326,7 +325,6 @@ type GetCartParam = {
     id?: string;
     i?: boolean;
     b?: boolean;
-    c?: boolean;
     assignCardId?: number;
     areaCode?: string;
     buyNow?: boolean;