+++ sdk/platform/Cart/CartPlatformApplicationValidator.d.ts
@@ -67,7 +67,6 @@ export = CartPlatformApplicationValidator;
  * @property {string} [id]
  * @property {boolean} [i]
  * @property {boolean} [b]
- * @property {boolean} [c]
  */
 /**
  * @typedef GetAddressByIdParam
@@ -522,7 +521,6 @@ type GetAbandonedCartDetailsParam = {
     id?: string;
     i?: boolean;
     b?: boolean;
-    c?: boolean;
 };
 type GetAddressByIdParam = {
     id: string;