+++ sdk/platform/Cart/CartPlatformApplicationValidator.js
@@ -83,7 +83,6 @@ const CartPlatformModel = require("./CartPlatformModel");
  * @property {string} [id]
  * @property {boolean} [i]
  * @property {boolean} [b]
- * @property {boolean} [c]
  */
 
 /**
@@ -513,7 +512,6 @@ class CartPlatformApplicationValidator {
       id: Joi.string().allow(""),
       i: Joi.boolean(),
       b: Joi.boolean(),
-      c: Joi.boolean(),
     }).required();
   }